+++ week2-task/js/dom.js
@@ -1,0 +1,15 @@
+export const tbody = document.getElementById("todoTableBody");
+export const topButtons = document.querySelectorAll(".filter__button");
+export const dropdownToggle = document.getElementById("dropdownToggle");
+export const priorityMenu = document.getElementById("priorityMenu");
+export const prioritySelectBtn = document.getElementById("prioritySelectBtn");
+export const prioritySelectMenu = document.querySelector(".main__top-menu");
+export const inputField = document.querySelector(".main__top-input");
+export const addButton = document.querySelectorAll(".main__top-button")[1];
+export const selectAllCheckbox = document.querySelector(
+  '.todo-table thead input[type="checkbox"]'
+);
+export const deleteButton = document.querySelectorAll(".main__middle-button")[0];
+export const completeButton = document.querySelectorAll(".main__middle-button")[1];
+export const completeModal = document.getElementById("completeModal");
+export const closeModalBtn = document.getElementById("closeModalBtn");
